Make some minor refactoring changes and move files out of plugins
For diff purposes
