--- UnityProject/Assets/Scripts/Plugins/SocketIOPlugin/SocketIOWrapper.cs.meta
@@ -1,13 +1,0 @@
-fileFormatVersion: 2
-guid: 5e1814392a1c047dd949f18662213fb9
-timeCreated: 1514977286
-licenseType: Free
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
