+++ UnityProject/Assets/Scripts/UnitySocketIO.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: 9822c85e79dcf442a87a7c003b928f60
+folderAsset: yes
+timeCreated: 1515153492
+licenseType: Free
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
